% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-distance.R
\name{get_distance}
\alias{get_distance}
\title{get_distance}
\usage{
get_distance(route_id, unit = "meters")
}
\arguments{
\item{route_id}{A trip's route ID (as a string).}

\item{unit}{The measurement unit for the distance (one of: "meters", "kilometers",
"feet", or "miles"). Defaults to "meters".}
}
\value{
The approximate distance covered by the trip (in the measurement unit specified).
}
\description{
Get trip distance
}
\details{
Get the distance covered by a trip, as identified through its route ID.
}
\examples{
\dontrun{
get_distance("route_04_2019@01253ad9-cfb8-44a5-9218-bff8109e4880")
}

}
