% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-duration.R
\name{get_duration}
\alias{get_duration}
\title{get_duration}
\usage{
get_duration(route_id, unit = "minutes")
}
\arguments{
\item{route_id}{A trip's route ID (as a string).}

\item{unit}{The measurement unit for the duration (one of: "seconds", "minutes", or "hours").
Defaults to "minutes".}
}
\value{
The approximate duration of the trip (in the measurement unit specified).
}
\description{
Get trip duration
}
\details{
Get the duration of a trip.
}
\examples{
\dontrun{
get_duration("route_04_2019@01253ad9-cfb8-44a5-9218-bff8109e4880")
}

}
