% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-summary-report.R
\name{get_summary_report}
\alias{get_summary_report}
\title{get_summary_report}
\usage{
get_summary_report(city, filename, path = ".", overwrite = FALSE)
}
\arguments{
\item{city}{The city for which to generate the summary report (as a string). Can be one of:
"Amherst", "Springfield", "Easthampton", "Northampton", "Holyoke", or "South Hadley".}

\item{filename}{The file name to use for the generated .pdf report.}

\item{path}{The path where to save the generated report.}

\item{overwrite}{Whether to overwrite any existing file with the same name at the specified
path. TRUE will overwrite existing files, FALSE will leave existing files
alone and issue an error. Defaults to FALSE.}
}
\description{
Get summary report
}
\details{
Generate a .pdf summary report for a city in which ValleyBike is active.
}
\examples{
\dontrun{
get_summary_report(city = "Amherst", filename = "amherst-report.pdf")
}

}
