% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-trajectory-data.R
\name{get_trajectory_data}
\alias{get_trajectory_data}
\title{get_trajectory_data}
\usage{
get_trajectory_data(route_id)
}
\arguments{
\item{route_id}{A trip's route ID (as a string).}
}
\value{
A data frame of trajectory data for the given route ID.
}
\description{
Get trajectory data
}
\details{
Get the point-by-point trajectory data for a trip, as identified through its route ID.
}
\examples{
\dontrun{
get_trajectory_data("route_07_2018@44bde968-5454-4db2-a4b1-a32493186f82")
}

}
