% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-trips.R
\name{identify_trips}
\alias{identify_trips}
\title{identify_trips}
\usage{
identify_trips(date = NULL, start_station = NULL, end_station = NULL)
}
\arguments{
\item{date}{The date that the trip was undertaken (as a YYYY-MM-DD string).}

\item{start_station}{The start station of the trip (as a string).}

\item{end_station}{The end station of the trip (as a string).}
}
\value{
A vector of route IDs that match the specified search criteria.
}
\description{
Identify trips
}
\details{
Identify the route IDs of specific trips using information about the date they were undertaken,
the start station, and/or the end station. One can provide only one or all of the arguments.
}
\examples{
\dontrun{
identify_trips(date = "2019-07-12", start_station = "Amherst Town Hall")
}

}
