% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-trip.R
\name{plot_trip}
\alias{plot_trip}
\title{plot_trip}
\usage{
plot_trip(route_id, api_key, layers, ...)
}
\arguments{
\item{route_id}{A trip's route ID (as a string).}

\item{api_key}{A Google Maps API key to pass to \code{ggmap::register_google} (required).}

\item{layers}{Any additional ggplot2 layers to add to the plot, such as titles, themes, etc.}

\item{...}{Arguments to pass to \code{ggmap::get_map} for customizing the map background,
such as \code{zoom}, \code{maptype}, \code{source}, \code{color}, etc.}
}
\value{
A ggplot2 plot object of the trip's trajectory on the map.
}
\description{
Plot trip
}
\details{
Plot the trajectory of a trip on the map.
}
\examples{
\dontrun{
plot_trip("route_04_2019@01253ad9-cfb8-44a5-9218-bff8109e4880", zoom = 16,
          layers = ggplot2::ggtitle("Trajectory of 01253ad9-cfb8-44a5-9218-bff8109e4880"))
}

}
