% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekend-or-holiday.R
\name{weekend_or_holiday}
\alias{weekend_or_holiday}
\title{weekend_or_holiday}
\usage{
weekend_or_holiday(route_id)
}
\arguments{
\item{route_id}{A trip's route ID (as a string).}
}
\value{
TRUE if the trip was undertaken on a weekend or federal holiday, FALSE otherwise.
}
\description{
Weekend or holiday
}
\details{
Determines whether a trip was undertaken on a weekend or on a federal holiday.
}
\examples{
\dontrun{
weekend_or_holiday("route_04_2019@01253ad9-cfb8-44a5-9218-bff8109e4880")
}

}
